V8 regexp.c: fix warnings in regdump
When compiling with DEBUG defined, two ptrdiff_t values were printed
as "%d" format, which resulted in a warning in clang (but not gcc).
This was an issue in 64 bit builds, as a pointer (diff) is 64 bit
while int is typically 32.

The most correct solution would be to use %td (ptrdiff_t format),
which is not supported in c89, but we have another correct solution
which does work in c89 or later:

Use %ld and cast the value to long - which is what this commit does.

This works great for us, because eventhough int and long are typically
the same size (32 bit), long in c89 and later guarantees 32 bit, while
int in c89 or later only guarantees 16 bits.

The actual value in practice is currently limited to 32K (regcomp
aborts if regsize >= 32767), so %ld with cast to long covers our
current needs, and still allows this limit to be raised up to 2M
without breaking the correctness of this code in c89 or later.

For what it's worth, regdump works and seems useful.

+++ regexp.c
@@ -1082,12 +1082,12 @@ regdump(regexp *r)
 	s = r->program + 1;
 	while (op != END) {	/* While that wasn't END last time... */
 		op = OP(s);
-		printf("%2d%s", s-r->program, regprop(s));	/* Where, what. */
+		printf("%2ld%s", (long)(s - r->program), regprop(s));
 		next = regnext(s);
 		if (next == NULL)		/* Next ptr. */
 			printf("(0)");
 		else 
-			printf("(%d)", (s-r->program)+(next-s));
+			printf("(%ld)", (long)((s - r->program)+(next-s)));
 		s += 3;
 		if (op == ANYOF || op == ANYBUT || op == EXACTLY) {
 			/* Literal string, where present. */
